\name{scale_baseline}
\alias{scale_baseline}
\title{scale for baseline data}
\usage{
scale_baseline()
}
\description{
  all u need to do is to put the root of document into the function
}
\examples{
## data sample please see sample_baseline
scale_baseline('C:/Users/Administrator/Desktop/fy_sample.xlsx')
}
